Rename and refactor c_compatibility_test.rs to secure_aggregation_test.rs
Fixes misleading test name and documentation that claimed C++ compatibility
but actually only tested secure aggregation functionality.

Changes:
- Rename file to reflect actual purpose (secure aggregation testing)
- Remove unused C++ test vector bytes
- Remove misleading documentation about C++ compatibility
- Clean up console output and debug artifacts
- Rename test functions to be more descriptive
- Focus tests on actual functionality being tested

The file now properly tests:
- Secure aggregation prevents rogue key attacks
- Normal aggregation fails verify_secure (security feature)
- Deterministic coefficient generation
- Key order independence

Real C++ compatibility testing is handled by true_cross_compatibility.rs.
