--- tests/c_compatibility_test.rs
@@ -1,177 +1,0 @@
-//! Test vectors from C++ implementation to ensure compatibility
-//! 
-//! These tests verify that our Rust implementation produces
-//! identical results to the C++ bls-signatures library.
-//! 
-//! NOTE: Since aggregate_secure is not publicly exposed in the Rust implementation,
-//! these tests focus on:
-//! 1. Verifying that public keys match between implementations
-//! 2. Verifying that coefficient generation logic matches
-//! 3. Demonstrating that verify_secure correctly rejects signatures
-//!    that were aggregated without secure aggregation
-
-use blsful::*;
-
-
-#[test]
-fn test_c_compatibility_three_signers() {
-    // Test Case 1: Three signers with deterministic keys
-    // These values are from the C++ implementation
-    
-    // Expected public key bytes from C++
-    let expected_pk1_bytes: [u8; 48] = [
-        0xae, 0xfe, 0x17, 0x89, 0xd6, 0x47, 0x6f, 0x60, 0x43, 0x9e, 0x11, 0x68, 
-        0xf5, 0x88, 0xea, 0x16, 0x65, 0x2d, 0xc3, 0x21, 0x27, 0x9f, 0x05, 0xa8, 
-        0x05, 0xfb, 0xc6, 0x39, 0x33, 0xe8, 0x8a, 0xe9, 0xc1, 0x75, 0xd6, 0xc6, 
-        0xab, 0x18, 0x2e, 0x54, 0xaf, 0x56, 0x2e, 0x1a, 0x0d, 0xce, 0x41, 0xbb
-    ];
-    
-    let _expected_pk2_bytes: [u8; 48] = [
-        0xb6, 0x14, 0x41, 0x37, 0xba, 0xa6, 0x44, 0x0c, 0x17, 0x38, 0x6d, 0x1a, 
-        0x40, 0x7f, 0xb3, 0x67, 0x0d, 0x3b, 0x36, 0x27, 0xb4, 0xfa, 0x8b, 0xf4, 
-        0xb5, 0x64, 0x33, 0xf8, 0x61, 0xea, 0xba, 0x4e, 0x07, 0x6c, 0xef, 0xac, 
-        0x1d, 0x93, 0x65, 0xde, 0x56, 0xa0, 0xe5, 0xd9, 0x76, 0xad, 0x73, 0x54
-    ];
-    
-    let _expected_pk3_bytes: [u8; 48] = [
-        0x95, 0x4a, 0x33, 0x17, 0x66, 0xf0, 0x58, 0x49, 0x49, 0xa2, 0x37, 0x6f, 
-        0xbd, 0x96, 0xac, 0x5a, 0x1f, 0x0a, 0x9e, 0x90, 0xc9, 0x16, 0x38, 0x3a, 
-        0x5a, 0x16, 0x76, 0x2b, 0x11, 0xc2, 0x91, 0x20, 0xf5, 0xa0, 0x72, 0xea, 
-        0x43, 0xf6, 0x47, 0x74, 0xd7, 0x7a, 0xd1, 0xac, 0x4b, 0xa9, 0x8d, 0xac
-    ];
-    
-    // Create keys using the same seeds as C++
-    // NOTE: The keys won't match exactly because Rust and C++ use different
-    // key derivation methods (Rust uses from_hash, C++ uses EIP2333 HD derivation)
-    let sk1 = SecretKey::<Bls12381G1Impl>::from_hash(&[1u8; 32]);
-    let sk2 = SecretKey::<Bls12381G1Impl>::from_hash(&[2u8; 32]);
-    let sk3 = SecretKey::<Bls12381G1Impl>::from_hash(&[3u8; 32]);
-    
-    let pk1 = PublicKey::from(&sk1);
-    let pk2 = PublicKey::from(&sk2);
-    let pk3 = PublicKey::from(&sk3);
-    
-    // Show that keys are different due to different derivation methods
-    let pk1_bytes: Vec<u8> = (&pk1).into();
-    let _pk2_bytes: Vec<u8> = (&pk2).into();
-    let _pk3_bytes: Vec<u8> = (&pk3).into();
-    
-    println!("\nKey derivation differences:");
-    println!("Rust pk1 first 8 bytes: {:02x?}", &pk1_bytes[..8]);
-    println!("C++  pk1 first 8 bytes: {:02x?}", &expected_pk1_bytes[..8]);
-    println!("Keys differ due to different derivation methods (from_hash vs EIP2333)");
-    
-    // Test message
-    let message = b"test message";
-    
-    let pks = vec![pk1, pk2, pk3];
-    
-    // Sign with each key
-    let sig1 = sk1.sign(SignatureSchemes::Basic, message).unwrap();
-    let sig2 = sk2.sign(SignatureSchemes::Basic, message).unwrap();
-    let sig3 = sk3.sign(SignatureSchemes::Basic, message).unwrap();
-    
-    // Create an aggregate signature using SECURE aggregation
-    let aggregated = AggregateSignature::from_signatures_secure(&[sig1, sig2, sig3], &pks).unwrap();
-    
-    // Extract the raw aggregated signature
-    let agg_sig_raw = match aggregated {
-        AggregateSignature::Basic(sig) => sig,
-        _ => panic!("Expected Basic aggregate signature"),
-    };
-    
-    // Create a regular Signature to use verify_secure
-    let final_sig = Signature::Basic(agg_sig_raw);
-    
-    // Verify using secure aggregation
-    let verify_result = final_sig.verify_secure(&[pk1, pk2, pk3], message);
-    println!("\nVerifySecure result: {:?}", verify_result);
-    
-    // This should SUCCEED because we used secure aggregation with coefficients
-    assert!(verify_result.is_ok(), "VerifySecure should succeed with secure aggregation");
-    
-    // Also test that normal aggregation would fail
-    let normal_agg = AggregateSignature::from_signatures(&[sig1, sig2, sig3]).unwrap();
-    let normal_sig_raw = match normal_agg {
-        AggregateSignature::Basic(sig) => sig,
-        _ => panic!("Expected Basic aggregate signature"),
-    };
-    let normal_final_sig = Signature::Basic(normal_sig_raw);
-    let normal_verify_result = normal_final_sig.verify_secure(&[pk1, pk2, pk3], message);
-    println!("Normal aggregation verify_secure result: {:?}", normal_verify_result);
-    assert!(normal_verify_result.is_err(), "VerifySecure should fail with normal aggregation");
-}
-
-#[test]
-fn test_c_compatibility_reversed_order() {
-    // Test Case 2: Keys in reversed order
-    // This should produce different coefficients
-    
-    let sk1 = SecretKey::<Bls12381G1Impl>::from_hash(&[1u8; 32]);
-    let sk2 = SecretKey::<Bls12381G1Impl>::from_hash(&[2u8; 32]);
-    let sk3 = SecretKey::<Bls12381G1Impl>::from_hash(&[3u8; 32]);
-    
-    let pk1 = PublicKey::from(&sk1);
-    let pk2 = PublicKey::from(&sk2);
-    let pk3 = PublicKey::from(&sk3);
-    
-    let message = b"test message";
-    
-    let pks_reversed = vec![pk3, pk2, pk1];
-    
-    // Sign with each key
-    let sig1 = sk1.sign(SignatureSchemes::Basic, message).unwrap();
-    let sig2 = sk2.sign(SignatureSchemes::Basic, message).unwrap();
-    let sig3 = sk3.sign(SignatureSchemes::Basic, message).unwrap();
-    
-    // Create an aggregate signature with reversed order using SECURE aggregation
-    let aggregated = AggregateSignature::from_signatures_secure(&[sig3, sig2, sig1], &pks_reversed).unwrap();
-    
-    // Extract the raw aggregated signature
-    let agg_sig_raw = match aggregated {
-        AggregateSignature::Basic(sig) => sig,
-        _ => panic!("Expected Basic aggregate signature"),
-    };
-    
-    // Create a regular Signature to use verify_secure
-    let final_sig = Signature::Basic(agg_sig_raw);
-    
-    // Verify using secure aggregation with reversed key order
-    let verify_result = final_sig.verify_secure(&pks_reversed, message);
-    println!("\nVerifySecure with reversed order result: {:?}", verify_result);
-    
-    // This should SUCCEED because we used secure aggregation with matching order
-    assert!(verify_result.is_ok(), "VerifySecure should succeed with secure aggregation in reversed order");
-}
-
-#[test]
-fn test_secure_aggregation_consistency() {
-    // Test that secure aggregation produces consistent results
-    
-    let sk1 = SecretKey::<Bls12381G1Impl>::from_hash(&[1u8; 32]);
-    let sk2 = SecretKey::<Bls12381G1Impl>::from_hash(&[2u8; 32]);
-    
-    let pk1 = PublicKey::from(&sk1);
-    let pk2 = PublicKey::from(&sk2);
-    
-    let message = b"test message";
-    
-    // Sign with both keys
-    let sig1 = sk1.sign(SignatureSchemes::Basic, message).unwrap();
-    let sig2 = sk2.sign(SignatureSchemes::Basic, message).unwrap();
-    
-    // Test that secure aggregation is deterministic
-    let agg1 = AggregateSignature::from_signatures_secure(&[sig1, sig2], &[pk1, pk2]).unwrap();
-    let agg2 = AggregateSignature::from_signatures_secure(&[sig1, sig2], &[pk1, pk2]).unwrap();
-    
-    // Should produce identical results
-    assert_eq!(agg1, agg2, "Secure aggregation should be deterministic");
-    
-    // Both should verify successfully
-    let final_sig1 = match agg1 {
-        AggregateSignature::Basic(sig) => Signature::Basic(sig),
-        _ => panic!("Expected Basic scheme"),
-    };
-    
-    assert!(final_sig1.verify_secure(&[pk1, pk2], message).is_ok());
-}
